requirements.txt
django-filebrowser<4
django-filter
django-grappelli<3
django-redis
django-rq==2.9.0
django==3.2.25
djangorestframework==3.15.1
drf-extensions

pillow
psycopg2-binary
drf_spectacular